+++ crates/testing-utils/src/lib.rs
@@ -1,22 +1,3 @@
-use std::{sync::mpsc, thread, time::Duration};
 pub mod bin;
 pub mod fs;
-
-pub fn panic_after<T, F>(d: Duration, f: F) -> T
-where
-    T: Send + 'static,
-    F: FnOnce() -> T,
-    F: Send + 'static,
-{
-    let (done_tx, done_rx) = mpsc::channel();
-    let handle = thread::spawn(move || {
-        let val = f();
-        done_tx.send(()).expect("Unable to send completion signal");
-        val
-    });
-
-    match done_rx.recv_timeout(d) {
-        Ok(_) => handle.join().expect("test panicked"),
-        Err(_) => panic!("test timeout"),
-    }
-}
+pub mod misc;
